refactor: 2차 머지 완료
