+++ linkup/src/main/java/com/learningcrew/linkup/common/dto/ApiResponse.java
@@ -36,7 +36,7 @@ public class ApiResponse<T> {
                 .message(message)
                 .build();
     }
-    public static <T> ApiResponse<T> fail(String message) {
+    public static <T> ApiResponse<T> failure(String message) {
         return ApiResponse.<T>builder()
                 .success(false)
                 .message(message)
