+++ linkup/src/main/java/com/learningcrew/linkup/linker/command/domain/aggregate/Member.java
@@ -5,11 +5,13 @@ import jakarta.persistence.*;
 import lombok.AccessLevel;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
+import org.apache.ibatis.type.Alias;
 
 import java.math.BigDecimal;
 import java.time.LocalDate;
 
-@Entity
+@Alias("UserMember")
+@Entity(name = "UserMember")
 @Table(name = "member")
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 @Getter
