+++ linkup/src/main/java/com/learningcrew/linkup/meeting/command/application/dto/request/MeetingCreateRequest.java
@@ -1,6 +1,6 @@
 package com.learningcrew.linkup.meeting.command.application.dto.request;
 
-import com.learningcrew.linkup.meeting.command.domain.aggregate.Gender;
+import com.learningcrew.linkup.common.constants.Gender;
 import jakarta.validation.constraints.Min;
 import jakarta.validation.constraints.NotBlank;
 import lombok.Getter;
