+++ linkup/src/main/java/com/learningcrew/linkup/meeting/command/domain/aggregate/Meeting.java
@@ -1,6 +1,7 @@
 package com.learningcrew.linkup.meeting.command.domain.aggregate;
 
 import com.fasterxml.jackson.annotation.JsonFormat;
+import com.learningcrew.linkup.common.constants.Gender;
 import jakarta.persistence.*;
 import lombok.AccessLevel;
 import lombok.Getter;
