+++ linkup/src/main/java/com/learningcrew/linkup/meeting/command/domain/aggregate/Member.java
@@ -4,8 +4,10 @@ import jakarta.persistence.*;
 import lombok.AccessLevel;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
+import org.apache.ibatis.type.Alias;
 
-@Entity
+@Alias("MeetingMember")
+@Entity(name = "MeetingMember")
 @Table(name = "member")
 @Getter
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
@@ -13,7 +15,7 @@ public class Member {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int member_id;
+    private int memberId;
 
     private String nickname;
 }
