+++ linkup/src/main/java/com/learningcrew/linkup/meeting/query/controller/MeetingParticipationQueryController.java
@@ -1,6 +1,6 @@
 package com.learningcrew.linkup.meeting.query.controller;
 
-import com.learningcrew.linkup.meeting.common.dto.ApiResponse;
+import com.learningcrew.linkup.common.dto.ApiResponse;
 import com.learningcrew.linkup.meeting.query.dto.response.MeetingParticipationResponse;
 import com.learningcrew.linkup.meeting.query.dto.response.ParticipantsResponse;
 import com.learningcrew.linkup.meeting.query.service.MeetingParticipationQueryService;
