+++ linkup/src/main/java/com/learningcrew/linkup/meeting/query/dto/request/MeetingSearchRequest.java
@@ -1,6 +1,6 @@
 package com.learningcrew.linkup.meeting.query.dto.request;
 
-import com.learningcrew.linkup.meeting.command.domain.aggregate.Gender;
+import com.learningcrew.linkup.common.constants.Gender;
 import jakarta.validation.constraints.NotEmpty;
 import jakarta.validation.constraints.NotNull;
 import lombok.Getter;
